+++ packages/web/src/components/controls/Switch/index.ts
@@ -1,7 +1,11 @@
 import { withStyleProps } from '../../../utils'
 import { StatefulSwitch } from './Components'
 import { getSwitchStyles } from './getStyles'
+import { SwitchProps, SwitchStyleProps } from './types'
 
-export const Switch = withStyleProps(StatefulSwitch, getSwitchStyles)
+export const Switch = withStyleProps<SwitchStyleProps, SwitchProps>(
+  StatefulSwitch,
+  getSwitchStyles,
+)
 
 export * from './types'