+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,34 @@
 # stories
 
+## 0.2.0
+
+### Minor Changes
+
+- Improves typing for web packages
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.30.0
+
+## 0.1.0
+
+### Minor Changes
+
+- Updating input types
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.29.0
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.28.0
+
 ## 0.0.4
 
 ### Patch Changes